feat: 댓글, 회고 기능 구현
- 댓글 생성, 수정, 삭제, 조회
- 회고 생성, 수정, 삭제, 조회

+++ src/main/java/com/gdgoc/study_group/round/application/CommentService.java
@@ -1,12 +1,12 @@
 package com.gdgoc.study_group.round.application;
 
 import com.gdgoc.study_group.comment.domain.Comment;
-import com.gdgoc.study_group.comment.dto.CommentDto;
 import com.gdgoc.study_group.exception.CustomException;
 import com.gdgoc.study_group.member.dao.MemberRepository;
 import com.gdgoc.study_group.member.domain.Member;
 import com.gdgoc.study_group.round.dao.RoundRepository;
 import com.gdgoc.study_group.round.domain.Round;
+import com.gdgoc.study_group.round.dto.CommentDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -28,14 +28,15 @@ public class CommentService {
      * 댓글을 작성합니다.
      *
      * @param roundId 댓글을 작성할 회차의 ID
-     * @param request 멤버 아이디와 댓글 내용
+     * @param memberId 댓글을 작성할 멤버의 ID
+     * @param request 댓글 내용
      * @return 생성한 댓글의 ID
      */
     @Transactional(readOnly = false)
-    public Long createComment(Long roundId, CommentDto request) {
+    public Long createComment(Long roundId, Long memberId, CommentDTO request) {
 
         Round round = roundRepository.findById(roundId).orElseThrow(() -> new CustomException(ROUND_NOT_FOUND));
-        Member member = memberRepository.findById(request.memberId()).orElseThrow(() -> new CustomException(MEMBER_NOT_FOUND));
+        Member member = memberRepository.findById(memberId).orElseThrow(() -> new CustomException(MEMBER_NOT_FOUND));
 
         Comment comment = Comment.create(round, member, request.comment());
         roundRepository.saveComment(comment);
@@ -48,14 +49,20 @@ public class CommentService {
      * 댓글을 수정합니다.
      *
      * @param roundId 수정할 댓글의 회차 ID
+     * @param memberId 댓글을 작성한 멤버의 ID
      * @param commentId 수정할 댓글의 ID
-     * @param updatedComment 수정할 댓글 내용
+     * @param request 수정할 댓글 내용
      */
     @Transactional(readOnly = false)
-    public void updateComment(Long roundId, Long commentId, String updatedComment) {
+    public void updateComment(Long roundId, Long memberId, Long commentId, CommentDTO request) {
 
-        Comment comment = roundRepository.findCommentByCommentId(roundId, commentId).orElseThrow(() -> new CustomException(COMMENT_NOT_FOUND));
-        comment.update(updatedComment);
+        Comment comment = roundRepository.findCommentByCommentId(roundId, commentId)
+                .orElseThrow(() -> new CustomException(COMMENT_NOT_FOUND));
+
+        System.out.println(comment.getMember().getId() + " AND " + memberId);
+        validateCommentOwner(comment, memberId);
+
+        comment.update(request);
     }
 
 
@@ -63,12 +70,16 @@ public class CommentService {
      * 댓글을 삭제합니다.
      *
      * @param roundId 삭제할 댓글의 회차 ID
+     * @param memberId 댓글을 작성한 멤버의 ID
      * @param commentId 삭제할 댓글의 ID
      */
     @Transactional(readOnly = false)
-    public void deleteComment(Long roundId, Long commentId) {
+    public void deleteComment(Long roundId, Long memberId, Long commentId) {
 
-        roundRepository.findCommentByCommentId(roundId, commentId).orElseThrow(() -> new CustomException(COMMENT_NOT_FOUND));
+        Comment comment = roundRepository.findCommentByCommentId(roundId, commentId)
+                .orElseThrow(() -> new CustomException(COMMENT_NOT_FOUND));
+        validateCommentOwner(comment, memberId);
+
         roundRepository.deleteCommentById(commentId);
     }
 
@@ -80,10 +91,10 @@ public class CommentService {
      * @param commentId 조회할 댓글의 ID
      * @return 댓글 내용
      */
-    public CommentDto getComment(Long roundId, Long commentId) {
+    public CommentDTO getComment(Long roundId, Long commentId) {
 
-        Comment comment = roundRepository.findCommentByCommentId(roundId, commentId).orElseThrow(() -> new CustomException(COMMENT_NOT_FOUND));
-        return CommentDto.from(comment);
+        return CommentDTO.from(roundRepository.findCommentByCommentId(roundId, commentId)
+                .orElseThrow(() -> new CustomException(COMMENT_NOT_FOUND)));
     }
 
 
@@ -93,9 +104,16 @@ public class CommentService {
      * @param roundId 댓글을 조회할 회차의 ID
      * @return 회차의 모든 댓글
      */
-    public List<CommentDto> getAllComments(Long roundId) {
+    public List<CommentDTO> getAllComments(Long roundId) {
 
-        roundRepository.findById(roundId).orElseThrow(() -> new CustomException(ROUND_NOT_FOUND));
-        return roundRepository.findComments(roundId).stream().map(CommentDto::from).toList();
+        return roundRepository.findComments(roundId).stream().map(CommentDTO::from).toList();
+    }
+
+
+    //==========HELPER==========//
+    private void validateCommentOwner(Comment comment, Long memberId) {
+        if (!comment.getMember().getId().equals(memberId)) {
+            throw new CustomException(FORBIDDEN);
+        }
     }
 }
